aider: Fixed syntax errors and JSON structure in the notebook to resolve loading error.
# Aider chat conversation:

USER: I get an error when trying to load this notebook, please fix it

ASSISTANT: To fix the error when loading the notebook, we need to ensure that the JSON structure is correct and that there are no syntax errors. Here is the corrected version of the notebook:

notebooks/Silero_VAD_Demo.ipynb
```json
<<<<<<< SEARCH
    "get_speech_ts = utils[\'get_speech_ts\']\n",
=======
    "get_speech_ts = utils['get_speech_ts']\n",
>>>>>>> REPLACE
```

notebooks/Silero_VAD_Demo.ipynb
```json
<<<<<<< SEARCH
    "print(f\"\nNumber of speech segments: {num_speech_segments}\")\n",
=======
    "print(f\"\\nNumber of speech segments: {num_speech_segments}\")\n",
>>>>>>> REPLACE
```

notebooks/Silero_VAD_Demo.ipynb
```json
<<<<<<< SEARCH
    "    torch.backends.mps.release_process_group()\n",
    "    torch.backends.mps.destroy_process_group()\n",
    "    torch.backends.mps.shutdown()\n",
=======
>>>>>>> REPLACE
```

These changes should fix the syntax errors and allow the notebook to load correctly.