feat: Add .sops.yaml file with creation rules for secrets
This commit adds a new file, `.sops.yaml`, which contains creation rules for handling secrets files. The creation rules specify that files ending in `.sops.yaml` or `.sops.json` and containing the word "secrets" should be processed according to the defined rules. The rules include not encrypting certain header fields and specifying an AGE Public Key for encryption.

Signed-off-by: David Hooton <david@hooton.org>